+++ src/Game/Functions.hpp
@@ -233,8 +233,8 @@ namespace Game
 	typedef __int64(*MSG_ReadInt64_t)(msg_t* msg);
 	extern MSG_ReadInt64_t MSG_ReadInt64;
 
-	typedef char*(*MSG_ReadString_t)(msg_t* msg, char* string, unsigned int maxChars);
-	extern MSG_ReadString_t MSG_ReadString;
+	typedef char*(*MSG_ReadBigString_t)(msg_t* msg);
+	extern MSG_ReadBigString_t MSG_ReadBigString;
 
 	typedef char*(*MSG_ReadStringLine_t)(msg_t *msg, char *string, unsigned int maxChars);
 	extern MSG_ReadStringLine_t MSG_ReadStringLine;